+++ tt/classes/com/oddlabs/tt/render/BillboardPainter.java
@@ -13,7 +13,9 @@ public abstract strictfp class BillboardPainter {
 	private final static FloatBuffer matrix_buf = BufferUtils.createFloatBuffer(16);
 	private final static DoubleBuffer plane_buf = BufferUtils.createDoubleBuffer(4);
 
-	private final static void initClipPlane(int clip_enum, int face_index, int vertex_index1, int vertex_index2, short[] indices, float[] face_tex_coords, float handedness) {
+	/* Team Penguin */
+	private final static void initClipPlane(int clip_enum, int face_index, int vertex_index1, int vertex_index2, int[] indices, float[] face_tex_coords, float handedness) {
+	/* End Penguin */
 		float u1 = getElement(face_index, vertex_index1, 0, 2, indices, face_tex_coords);
 		float v1 = getElement(face_index, vertex_index1, 1, 2, indices, face_tex_coords);
 		float u2 = getElement(face_index, vertex_index2, 0, 2, indices, face_tex_coords);
@@ -42,12 +44,14 @@ public abstract strictfp class BillboardPainter {
 		GL11.glEnable(GL11.GL_CLIP_PLANE2);
 	}
 
-	private final static float getElement(int face_index, int vertex_index, int element_index, int vertex_size, short[] indices, float[] vertices) {
+	/* Team Penguin */
+	private final static float getElement(int face_index, int vertex_index, int element_index, int vertex_size, int[] indices, float[] vertices) {
 		int vertices_index = indices[face_index*3 + vertex_index];
 		return vertices[vertices_index*vertex_size + element_index];
 	}
 
-	public final static void loadFaceMatrixAndClipPlanes(int face_index, short[] indices, float[] face_vertices, float[] face_tex_coords) {
+	public final static void loadFaceMatrixAndClipPlanes(int face_index, int[] indices, float[] face_vertices, float[] face_tex_coords) {
+	/* End Penguin */
 		// Find object space to texture space matrix, mapping vectors in object space to vectors in texture space
 		Vector3f v1 = new Vector3f(getElement(face_index, 0, 0, 3, indices, face_vertices),
 								   getElement(face_index, 0, 1, 3, indices, face_vertices),
